Implement CRUD of art toy review

+++ api/infrastructure/handler/arttoy/getarttoybyid.go
@@ -12,7 +12,7 @@ import (
 )
 
 type GetArtToyByIDInput struct {
-	ID int `path:"id"`
+	ID int64 `path:"id"`
 }
 
 type GetArtToyByIDOutput struct {
